requirements.txt
djangorestframework==3.11.2
Django==3.2
gunicorn==20.0.4
PyJWT==1.7.1
djangorestframework-jwt
requests
djangorestframework-simplejwt
django-cors-headers
django-dotenv
numpy
pillow
html5lib
Drf-writable-nested
importlib-resources
